+++ client/client.go
@@ -13,13 +13,9 @@ import (
 	"golang.org/x/time/rate"
 )
 
-type ctxKey string
-
 const (
 	BurstLimit        = 10
 	RequestsPerSecond = 10
-
-	clientConextKey ctxKey = "client"
 )
 
 var (
@@ -62,24 +58,4 @@ func New(ctx context.Context, tkn string) (*Client, error) {
 		ghClient: github.NewClient(oauth2.NewClient(ctx, ts)),
 		rate:     rl,
 	}, nil
-}
-
-func WithClient(ctx context.Context, tkn string) context.Context {
-	ctx, cancel := context.WithCancelCause(ctx)
-
-	c, err := New(ctx, tkn)
-	if err != nil {
-		cancel(err)
-	}
-
-	return context.WithValue(ctx, clientConextKey, c)
-}
-
-func ClientFromContext(ctx context.Context) (*Client, error) {
-	c, ok := ctx.Value(clientConextKey).(*Client)
-	if !ok {
-		return nil, ErrClientNotFound
-	}
-
-	return c, nil
 }
